feat(draggableTabs): Add new draggable tabs component  (#75025)
This PR creates the `<DraggableTabs/>` component along with many
supporting components. It utilizes the new tab variant [recently
introduced](https://github.com/getsentry/sentry/pull/74867), and adds
support for reordering them via dragging (implemented w/ Framer motion).

This PR contains a subset of changes from a previous
[PR](https://github.com/getsentry/sentry/pull/73543) that contained a
lot of new changes to the draggable tabs component, and it will be the
first in a series of PRs that break up that PR into smaller chunks. This
PR _should_ be the largest of them all, since it creates all of the
draggable tabs components. I have tried to pare the old PR down to
_only_ contain the changes necessary to get the draggable tabs to be
actually draggable (using Framer motion), but please feel free to
comment on anything that looks like it may be an artifact of these
changes.

Known issues in this current implementation that will be fixed in
subsequent PRs:

- There are no tests for this component 
- Overflow tabs interact with the tab dividers very weirdly 
- Tab names wrap very weirdly 


https://github.com/user-attachments/assets/329fe197-b6af-47af-8faa-4d34ef65181f