feat(draggableTabs): Add query count and tab menu button  (#75114)
closes #73224

This PR adds the query count and tab dropdown and button to the
draggable tabs component. These changes are purely visual, no
functionality for the dropdown menu options have been implemented.

<img width="460" alt="image"
src="https://github.com/user-attachments/assets/b22ff7d8-be53-432d-8232-a12f5db7709e">

As an added bonus, I also have cleaned up the DraggableTabList and
DraggableTab components so that they are a lot more generic than the
previous implementation of this feature

Known Issues:
- Dragging a tab to the left makes it disappear into the overflow menu
for some reason
- Hovering over the elements in the dropdown menu causes the tab itself
to increase in opacity as if its being hovered over. Clicking on an
option in the dropdown menu also increases the opacity of the tab