ref(seer grouping): Handle Seer only sending matching groups (#75209)
Seer no longer sends back similar but non-matching groups when called during ingest, which means on the Sentry side, we no longer need to handle that case. This PR therefore removes the `should_group` check on retrieving a parent group, since Seer will only send back results if `should_group` is true. It also stops storing the Seer results on the group, since a) the only time there's a new group on which to store the results is when a match isn't found, and b) in cases where a match isn't found, there now won't be any similar groups to store, only an empty results list. Finally, since the metadata is now only going to be stored on the event, and the event knows its own hash, `request_hash` has been removed from the metadata we store.