ref(controlsilo) Add forwards compatibility for integration tasks (#75363)
Add forwards compatible tasks for github, jira and opsgenie plugins, so
that when we update callers we won't drop any tasks. These aliases will
also ensure continuity should we need to rollback.

The next step of this work is to move the implementation of these tasks
into the new modules and have the old modules call the new ones. After
that is complete, we'll remove the old module locations.

Split off of #74925