feat(breadcrumbs): Add level filtering and previous search logic (#75536)
Tried to simplify search but it might be behaving differently in a way I
haven't been able to test. Since there was really nothing wrong with the
old one, I've made it work for the new UI and removed 'new search' which
wasn't markedly better (maybe faster? but I'm not sure).

Also added back level filtering, which was removed when we had decided
to omit levels in a previous iteration. It was mistakenly not added back
then.

![image](https://github.com/user-attachments/assets/1bcdc003-6b8c-4e12-b9b5-e3829e295971)