chore(seer grouping): Remove redundant metrics (#75565)
We have a bunch of metrics which were added before the `did_call_seer` metric was added, and which are now duplicated by instances of that metric with tags corresponding to the metrics in question. (For example, where we've been collecting a `grouping.similarity.seer_global_killswitch_enabled` metric, we've also been collecting the `did_call_seer` metric with the tag `blocker: global-killswitch`.) This removes all of the redundant metrics (which, it should be noted, we also haven't even been using).