ref(seer grouping): Use option to control sample rate of metrics (#75564)
Right now, while Seer grouping is still in EA, the number of times our various metrics are hit is relatively small. For that reason, we've kicked the sample rate on many up them up to 100%. In order to standardize this (and to be able to change it easly), this PR uses a new option, `seer.similarity.metrics_sample_rate`, to control the sample rate for all of the metrics on our dashboard.

Note: Checking an option in the `@metrics.wraps` calls in `tasks.embeddings_grouping.utils` caused tests to fail in about a dozen different test files (most of them unrelated, weirdly) with a complaint that database access isn't allowed unless marked in the test file. Rather than spend a ton of time chasing that down, I just converted the `@metrics.wraps` calls to `with metrics.timer` calls (which is all `@metrics.wraps` does), which is why the changes seem more extensive in that file than the others.