ref(compact-select): Use key instead of value for hidden/disabled options (#75751)
CompactSelect uses `hiddenOptions` and `disabledOptions` to determine
which options are hidden or disabled. Currently, they track the option
`value`, not the option `key`. For most use cases, the key is an escaped
version of the value.

I want to change this to track the `key` for a couple reasons:
1. Because keys are guaranteed to be unique, it makes more sense to use
that instead of the value
2. For the search bar, I want options that have the same `value`, but
different `key`s, which is not possible with the current implementation
of these utils.