fix(dashboard): Do not auto-resize for builder and add to dashboard (#75908)
The only place where we want widgets to be resizable are in custom
dashboards. Since we added the ID to widget builder state, the
auto-resizing was triggered but there is no container with fixed height
to resize properly.

Before we were relying on context about the widget state, but that was a
leaky abstraction. It's just better to be explicit here.