ref(integration): Introduce MessagingIntegrationSpec (#74743)
Begin unifying the three messaging-related integrations (Slack, MS
Teams, and Discord) with an abstract base class. The base class is a
prototype that extracts duplicated code in the short term, and in the
long term may serve as a comprehensive "feature inventory" of all
Sentry's features related to messaging integrations.