fix(seer grouping): Handle Seer returning grouphash with no group (#75861)
We have lately run into [an issue[1] where Seer is recommending hashes with no group attached to them, which has highlighted a) the fact that this is even a possibility, and b) the fact that though we've been talking about Seer sometimes recommending non-existent _groups_, in fact all we check for is the existence of the grouphash. (The assumption there being that all grouphashes point to a group. If everything else is working correctly that should be true, but clearly stuff happens.)

This PR therefore makes two changes:

- Change the `SimilarGroupNotFoundError` to be a `SimilarHashNotFoundError`, to more accurately reflect the situation when it arises.

- Add a new `SimilarHashMissingGroupError` to denote the case described above, so that we can track it rather than silently swallowing the problem.


[1] https://github.com/getsentry/sentry/issues/75860