chore(custom-views): Make id required in group serach view response type (#75988)
This PR makes a minor update to the frontend functions that call the
groupsearchviews endpoints. Specifically - it makes the id parameter
required in the _response_ type of both the GET and PUT, but leaves it
optional in the _request_ type. This did not cause any issues before,
but is more correct and will help with the existing custom views work.