feat(query-builder): Add defaultValue and values to the field definition (#76083)
Allows the default value to be overridden in the field definition.
Useful if `100` or `10ms` isn't in the right ballpark. It is already
useful for the `is` filter which had a hardcoded `unresolved` as the
default value.

Also adds `values` to the field definition. Currently that value is
stored in the `Tag`, but I am planning on moving all data to the field
definition rather than maintaining the config in two separate places.
For now, `values` in the `Tag` overrides anything in the field
definition.