ref(issue-details): Remove <GroupEventEntry /> (#76078)
This PR removes `GroupEventEntry` from the issue details page, and
renames its component `EventDetailsContent` rather than
`DefaultGroupEventDetailsContent`.

The reasons for this are a few:
- I think we should get rid of `EventEntry` wholesale. It is a component
which can render almost 13 different components based on its props which
makes it difficult to work with as you have to assume a lot of tenuously
linked components have similar styles and layouts.
- `EventEntry` is used in `EventEntries` and `Entries` and another newer
`Entries`, all components that I think should be removed as well. They
seem to have the goal of displaying all the event data that you may find
on issue details, but they are similar at best. It is not in parity with
issue details and updates do not always make their way to issue details
or vice versa.

I think the next steps are to replace the above 3 components with direct
calls to `EventDetailsContent` and maybe add some config for disabling
sections that the caller does not want rendered, allowing each usage to
be explicit about what gets shown. Currently, if you were to add to
`EventEntry`, the changes would be propagated to discover, trace view,
shared issues and not issue details, since this system is half adopted.