fix(discord): Fix Bot Command Timeout (#76197)
Fixes
https://sentry.sentry.io/issues/5131244670/events/a169bd4760e9408f86f40822285b804f/
& Discord Bot Commands that are currently failing:

![image](https://github.com/user-attachments/assets/9ef653d3-72f1-4ec2-a679-a7d148518128)

What went wrong?

Discord expects us to respond to a webhook within 3 seconds, which is
difficult to do when we are making RPC calls between region and control.

Currently the way handle the latency induced by RPC calls is to send a
discord a response to a webhook, which tells it that we are processing.
this buys us 15 minutes to respond with the new data, which is more
enough.

However, we didn't follow the Discord API spec correctly, and out
response was not formatted & we sent the incorrect status code.
> When responding to an interaction received via webhook, your server
can simply respond to the received POST request. You'll want to respond
with a 200 status code (if everything went well), as well as specifying
a type and data....

(https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-response-object-messages)

What this meant is in reality, we only had 3 seconds to respond and when
the RPC call didn't finish in time, Discord believed we didn't respond.
In this pr, i actually send the correct response, which ultimately gives
us the full 15 minutes.

In the video below, i exaggerated how long it would take to respond to
show the functionality. Notice how it says, the bot is thinking, which
shows we told Discord that we are working on a response.


https://github.com/user-attachments/assets/ae7ad9bb-8ff2-4c34-bb01-30d8d22c1190


Note: We run through this async logic only when running on silo mode,
which means testing this locally on monolith won't surface this issue.